refac(rpc): extract function
