+++ crates/rpc/src/v03/method/estimate_fee.rs
@@ -1,11 +1,10 @@
 use crate::{
-    cairo::ext_py::{BlockHashNumberOrLatest, GasPriceSource},
     context::RpcContext,
     v02::types::{reply::FeeEstimate, request::BroadcastedTransaction},
 };
-use pathfinder_common::{BlockId, StarknetBlockTimestamp};
-use starknet_gateway_types::pending::PendingData;
-use std::sync::Arc;
+use pathfinder_common::BlockId;
+
+use super::common::prepare_handle_and_block;
 
 #[derive(serde::Deserialize, Debug, PartialEq, Eq)]
 pub struct EstimateFeeInput {
@@ -39,34 +38,8 @@ pub async fn estimate_fee(
     context: RpcContext,
     input: EstimateFeeInput,
 ) -> Result<Vec<FeeEstimate>, EstimateFeeError> {
-    let handle = context
-        .call_handle
-        .as_ref()
-        .ok_or_else(|| anyhow::anyhow!("Unsupported configuration"))?;
-
-    // discussed during estimateFee work: when user is requesting using block_hash use the
-    // gasPrice from the starknet_blocks::gas_price column, otherwise (tags) get the latest
-    // eth_gasPrice.
-    //
-    // the fact that [`base_block_and_pending_for_call`] transforms pending cases to use
-    // actual parent blocks by hash is an internal transformation we do for correctness,
-    // unrelated to this consideration.
-    let gas_price = if matches!(input.block_id, BlockId::Pending | BlockId::Latest) {
-        let gas_price = match context.eth_gas_price.as_ref() {
-            Some(cached) => cached.get().await,
-            None => None,
-        };
-
-        let gas_price =
-            gas_price.ok_or_else(|| anyhow::anyhow!("Current eth_gasPrice is unavailable"))?;
-
-        GasPriceSource::Current(gas_price)
-    } else {
-        GasPriceSource::PastBlock
-    };
-
-    let (when, pending_timestamp, pending_update) =
-        base_block_and_pending_for_call(input.block_id, &context.pending_data).await?;
+    let (handle, gas_price, when, pending_timestamp, pending_update) =
+        prepare_handle_and_block(&context, input.block_id).await?;
 
     let result = handle
         .estimate_fee(
@@ -79,52 +52,6 @@ pub async fn estimate_fee(
         .await?;
 
     Ok(result)
-}
-
-/// Transforms the request to call or estimate fee at some point in time to the type expected
-/// by [`crate::cairo::ext_py`] with the optional, latest pending data.
-async fn base_block_and_pending_for_call(
-    at_block: BlockId,
-    pending_data: &Option<PendingData>,
-) -> Result<
-    (
-        BlockHashNumberOrLatest,
-        Option<StarknetBlockTimestamp>,
-        Option<Arc<starknet_gateway_types::reply::PendingStateUpdate>>,
-    ),
-    anyhow::Error,
-> {
-    use crate::cairo::ext_py::Pending;
-
-    match BlockHashNumberOrLatest::try_from(at_block) {
-        Ok(when) => Ok((when, None, None)),
-        Err(Pending) => {
-            // we must have pending_data configured for pending requests, otherwise we fail
-            // fast.
-            match pending_data {
-                Some(pending) => {
-                    // call on this particular parent block hash; if it's not found at query time over
-                    // at python, it should fall back to latest and **disregard** the pending data.
-                    let pending_on_top_of_a_block = pending
-                        .state_update_on_parent_block()
-                        .await
-                        .map(|(parent_block, timestamp, data)| {
-                            (parent_block.into(), Some(timestamp), Some(data))
-                        });
-
-                    // if there is no pending data available, just execute on whatever latest.
-                    Ok(pending_on_top_of_a_block.unwrap_or((
-                        BlockHashNumberOrLatest::Latest,
-                        None,
-                        None,
-                    )))
-                }
-                None => Err(anyhow::anyhow!(
-                    "Pending data not supported in this configuration"
-                )),
-            }
-        }
-    }
 }
 
 #[cfg(test)]
@@ -223,6 +150,7 @@ mod tests {
     // These tests require a Python environment properly set up _and_ a mainnet database with the first six blocks.
     mod ext_py {
         use std::path::PathBuf;
+        use std::sync::Arc;
 
         use super::*;
         use crate::v02::types::request::{
